Step through bugs in file association
