+++ public/components/appManager.js
@@ -37,15 +37,14 @@ class AppManager {
   // Check process.argv after startup (win)
   static checkAndOpenFileFromArgs() {
     log.info('checkAndOpenFileFromArgs', process.argv);
-    if (process.platform === 'win32') {
+    if (process.platform === 'win32' || process.platform === 'linux') {
       AppManager.openFileFromArgs(process.argv);
     }
   }
 
   static openFileFromArgs(argv) {
     log.info('openFileFromArgs', argv);
-
-    if (process.platform === 'win32') {
+    if (process.platform === 'win32' || process.platform === 'linux') {
       const filePath = getFileFromArgs(argv);
 
       if (filePath) {
